pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "aisafeguard"
version = "0.1.0"
description = "Safety rails for every AI app — model-agnostic guardrails for LLM applications"
readme = "README.md"
license = {text = "Apache-2.0"}
requires-python = ">=3.10"
authors = [
    { name = "AISafe Guard Contributors" },
]
keywords = ["ai", "safety", "guardrails", "llm", "security", "prompt-injection", "pii"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: Apache Software License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Topic :: Security",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Typing :: Typed",
]
dependencies = [
    "pydantic>=2.0",
    "pyyaml>=6.0",
    "click>=8.0",
    "anyio>=3.7,<5",
]

[project.optional-dependencies]
ml = [
    "transformers>=4.30",
    "torch>=2.0",
    "detoxify>=0.5",
    "sentence-transformers>=2.0",
]
proxy = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.20",
    "httpx>=0.24",
]
integrations = [
    "openai>=1.0",
    "anthropic>=0.20",
    "langchain-core>=0.1",
    "litellm>=1.0",
]
telemetry = [
    "opentelemetry-api>=1.20",
    "opentelemetry-sdk>=1.20",
]
all = [
    "aisafeguard[ml,proxy,integrations,telemetry]",
]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "ruff>=0.4",
    "mypy>=1.8",
]

[project.scripts]
aisafe = "aisafeguard.cli.main:cli"

[project.urls]
Homepage = "https://github.com/aisafeguard/aisafeguard"
Documentation = "https://aisafeguard.dev"
Repository = "https://github.com/aisafeguard/aisafeguard"
Issues = "https://github.com/aisafeguard/aisafeguard/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.10"
strict = true
